feat: Creation of a custom hook for authorization with GitHub.
